widget: [API] correct default scaling of images
When no scale factor is set, scale by 1.0, mapping one image pixel to
one device-independent pixel. This matches the behavior of CSS and other
frameworks.

The old code attempted to convert to Dp while taking the image's DPI
into account. This was wrong in two ways:

- It assumed that the default display DPI is 160, but this is only true
  for Android. Other platforms use 96, 162, or leave it undefined. Thus
  image.Layout's idea of a dp didn't match that of Gio on most
  platforms.

- It tried to account for image DPI, and assumed a default of 72. This
  was wrong in that DPI in images is merely metadata meant for printing,
  not display. The vast majority of software such as image viewers and
  image editors do not take DPI into account, mapping one image pixel
  either to one physical pixel or to one device-independent pixel. That
  is, users would expect their images to either display 1 to 1, or scaled
  based on PxPerDp, but not scaled based on the image's DPI.

We default to a scale of 1 to stay consistent with other parts of Gio
that scale by default. Users who don't want any scaling can continue to
set the scale to the inverse of PxPerDp.

While we're here we clarify the documentation of the Scale field.

This change is backwards incompatible for users that relied on the
default scale.

Signed-off-by: Dominik Honnef <dominik@honnef.co>