widget: update textIterator docs for accuracy
The previous docs claimed that failing to set a textMaterial would result in
invisible glyphs when in reality it results in using whatever the current paint
material is. This could be the paint material from before laying out the glyphs,
or it could be the material for a bitmap glyph. As such, it's better to say that
the color is undefined.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>