text: handle shaping string containing only newline
This commit ensures that we properly handle the case in which an input string is only
a newline character. We now make a run of text by shaping a space rune and then
drop the glyph/rune data from the space (keeping the line height and such).
The prior behavior would shape zero runes, resulting in no output runs, and thus our
logic for synthesizing a glyph for the newline would never execute while iterating the
runs.

I tried to restructure to instead catch whether there were zero runs after the iteration,
but it came out much uglier and harder to understand that way.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>