text: fix additional truncated newline bug
This commit fixes another rune accounting bug that the fuzzer discovered. If we
shaped a space in order to acquire line metrics, but the space itself was truncated,
we would reset the truncated count to zero. This had the side effect of lying to
later logic about whether the truncator run was present at the end of the shaped
text.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>