io/transfer,app: [macOS] add support for file opening events
Correctly handle macOS file opening events so that an app containing a
binary using gio can be used to open certain files and even be set as
default app for certain file extensions.
This is done via implementing application:openFile in GioAppDelegate.

File events issued this way will be made available to a gio through the
io/transfer package. This package now also contains documentation for
how to register a macOS app as able to support certain file extensions.