app: raiseWindow on macos should activate the app if not active
Calling window.Perform(system.ActionRaise) does not show the window on the top if the app is currently not active. This can happen for example, if the app integrated with systray (https://pkg.go.dev/fyne.io/systray) where the menu item launches a window, the window is not showing at the top. It is fixed by activating the current app if necessary

Signed-off-by: Siva Dirisala <siva.dirisala@gmail.com>