app,io,widget: [android/macos] fix focus issues
This patch is intended to fix two issues related to focus and
software-keyboard.

The focus is now reseted if it's lost due to external event,
such as clicking on non-Gio window/activity/fragment.

The software-keyboard will now re-open when clicked, except
if it's ReadOnly.

On macOS it will request focus again if the focus is lose
due to another View in the same Window. However, currently,
it's done by using ShowTextInput.

Fixes: https://todo.sr.ht/~eliasnaur/gio/591

Signed-off-by: inkeliz <inkeliz@inkeliz.com>